Merge raygui branch with main branch

+++ include/midi.h
@@ -11,7 +11,8 @@
  * Change the ADSR parameters when the assigned knobs are being triggered
  * Change the cutoff, detune and amplification when the assigned knobs are being triggered
  */
-int get_midi(snd_rawmidi_t *midi_in, synth_t *synth,
-             double *attack, double *decay, double *sustain, double *release);
+int 
+get_midi(snd_rawmidi_t *midi_in, synth_t *synth,
+             float *attack, float *decay, float *sustain, float *release);
 
 #endif
